week_2_orchestration/requirements.txt
psycopg2-binary==2.9.9
apache-airflow-providers-amazon==9.2.0
apache-airflow-providers-postgres==6.0.0